Reader, Dan comment consideration.
Did some more work to the reader which is sort of working not really. Something to do with how I am instantiating the polygon and putting it in the list not really with how I'm reading.

Got rid of MyPolgons ability to return nodes. It should only be able to return Polygons that are stored inside nodes. This is based on comments on PA1 from Dan.

Peace out